Replaced deprecated method.
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1540418 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/math3/util/Combinations.java
@@ -399,7 +399,7 @@ public class Combinations implements Iterable<int[]> {
                     throw new OutOfRangeException(digit, 0, n - 1);
                 }
 
-                ret += c[i] * ArithmeticUtils.pow(n, (long) i);
+                ret += c[i] * ArithmeticUtils.pow(n, i);
             }
             return ret;
         }
